requirements.txt
httpx
groq
python-dotenv
onepassword-sdk
